Third Year Project - CIT Mobile Development
Overview

Student populations, by their very nature, are mobile and there is a
complete turnover in the student population every 4 years

(approximately). CIT is no different in this regard.  Every September,
thousands of Fresher and Erasmus students have their first

experience of the college, very often with little more than a handbook
to guide them. In addition to these new students, existing

students are often not much better informed having suffered serious
memory losses over the summer months.

Your brief, for the 3rd year project, is to design an app “Virtual CIT”,
which enables the following features:

1) Provide an App interface to the myCIT portal.

2) Information portal for student union and student services. For
example, all information in the student handbook should be

available within the app.

3) Integration with Social media such as the College’s twitter and
Facebook feeds.

4) Viewing of existing timetables in a mobile friendly format.

5) Virtual College Tour (VCT) using Google “Tour Builder”.

6) Location of lecture rooms and labs (Location Feature).

7) Custom designed feature unique to each group (each group implements a
unique feature not already outlined in the spec).

One of the key objectives of the project is to implement the Android
automation of finding lecture rooms and labs in the college. It

is envisaged that a key component of the College app is to integrate
with Goggle Maps via its INDOOR Maps API. This will

implement the graphical element of the “Location Feature” within the
App.

The choice of technology used in the implementation of the Location
Feature will be decided by each group. Navigation could be

implemented using the device’s GPS, the data plan of the device or the
app could be designed to be standalone i.e. once it is

downloaded, there is no further need for downloading data once the app
is launched.

The app will be designed to be generic so that it may easily be
customised for any College. It will also be possible for a website

administrator to add features and update content on the app via a
website backend.

The app should also be scalable i.e. it should enable an admin user to
create other entries for other Colleges. The information on

this website, and its associated MySQL database, will act as an
interface for adding future colleges to the App.

The site will be implemented using a PHP website with a Joomla CMS (or
another CMS of your choice). Students may use a

package such as XAMPP as the development environment. The database for
the backend will be implemented in a DBMS such as

mySQL.

Once the App is fully developed, it is hoped to place it on the Android
market place, as a free download. The App should be

designed so that it can be used in several languages. For the purposes
of this project, it is only necessary to implement the App in

English. However, the App should be designed to be scalable so that it
could be implemented in another European language at a later stage.